# Optimize removeDuplicateGradient to avoid O(n²) in the average case