@nthykier 2018-03-10
Ideally the attributed picked in computeGradientBucketKey would be carefully selected to be (in order):

 * most likely to be unique/different when the gradients are different
 * combined generate the smallest possible key

I do not know what most of the attributes do, so I cannot claim that I have picked the best possible attributes for this purpose.  If there are attributes better suited for this purpose, then please let me know and I will update the patch to use those.