@nthykier 2018-03-10
I am retracting this pull request as this problem is better solved by going "all-in" on the bucket'ing mechanism.  The MediaWiki_logo_1.svg file contains over 900 duplicated radialGradients, so even with this pull request, this logo triggers a bad case of O(n²).